Start tests after server responds
It seems that 15ms sleep was not enough sometimes, causing tests to
fail, but on the other hand locally at least was more than enough time
to wait for sever to start up.

Changed the tests setup to check that server actually responds, before
starting the tests. This way tests should be faster, and not
intermittently fail.